requirements.txt
numpy
torch
e3nn
mace-torch
pytorch-lightning
omegaconf
hydra-core
tensorboard
torchmetrics
wandb
plotly
matplotlib
nbformat
scikit-learn
scipy
deeptime
dscribe
ase
hdbscan
seaborn
umap-learn
pytest
geomloss
